refactor main page
